// sigpcb-backoffice-frontend/src/app/core/interfaces/role.ts
export type PermissionExpression =
  | 'all'
  | 'read-admin'
  | 'edit-admin'
  | 'read-roles'
  | 'edit-roles'
  | 'read-signatures'
  | 'edit-signatures'
  | 'read-qcm-management'
  | 'edit-qcm-management'
  | 'read-inspections'
  | 'edit-inspections'
  | 'read-dossier-validation'
  | 'edit-dossier-validation'
  | 'read-programming'
  | 'edit-programming'
  | 'read-licences-management'
  | 'edit-licences-management'
  | 'read-permit-authentication'
  | 'edit-permit-authentication'
  | 'read-permit-replacement-duplicate'
  | 'edit-permit-replacement-duplicate'
  | 'read-international-permit'
  | 'edit-international-permit'
  | 'read-log'
  | 'read-permit-exchange'
  | 'edit-permit-exchange'
  | 'read-permit-extension'
  | 'edit-permit-extension'
  | 'read-statistics'
  | 'read-faq-management'
  | 'edit-faq-management'
  | 'read-examiner-recruitment'
  | 'edit-examiner-recruitment'
  | 'read-monitor-recruitment'
  | 'edit-monitor-recruitment'
  | 'read-company-recruitment-management'
  | 'edit-company-recruitment-management'
  | 'read-composition-recruitment-management'
  | 'edit-composition-recruitment-management'
  | 'read-exam-results-recruitment'
  | 'edit-exam-results-recruitment'
  | 'read-customer-support'
  | 'edit-customer-support'
  | 'read-agenda-management'
  | 'edit-agenda-management'
  | 'read-annex-management'
  | 'edit-annex-management'
  | 'read-permit-categories-management'
  | 'edit-permit-categories-management'
  | 'read-languages-management'
  | 'edit-languages-management'
  | 'read-restrictions-management'
  | 'edit-restrictions-management'
  | 'read-reporting-management'
  | 'read-exam-results-management'
  | 'read-driving-school-management'
  | 'edit-driving-school-management'
  | 'read-agreement-management'
  | 'edit-agreement-management'
  | 'read-base-configuration'
  | 'edit-base-configuration'
  | 'read-titles-management'
  | 'edit-titles-management'
  | 'read-administrative-units-management'
  | 'edit-administrative-units-management'
  | 'read-candidates-menu-access'
  | 'read-transmission-menu-access'
  | 'read-chapitres-management'
  | 'read-department'
  | 'read-commune'
  | 'read-arrondissement'
  | 'read-qestion-management'
  | 'read-reponses-possible-management'
  | 'read-baremes-management'
  | 'read-dispense-paiement'
  | 'edit-dispense-paiement'
  | 'manage-dispense-paiement';

export interface Permission {
  id: number;
  name: PermissionExpression;
  nom_complet: string;
  description: string;
}
export interface Role {
  id: number;
  name: string;
  nom_complet: string;
  permissions: Permission[];
}